Bugfix remove path query from all media pages use query params
